+++ qr_code/views.py
@@ -47,17 +47,17 @@ def serve_qr_code_image(request) -> HttpResponse:
     # Handle image access protection (we do not allow external requests for anyone).
     check_image_access_permission(request, qr_code_options)
     try:
-        text = base64.b64decode(request.GET.get('text', '')).decode(encoding='utf-8')
+        data = base64.b64decode(request.GET.get('data', '')).decode('utf-8')
     except binascii.Error:
         raise SuspiciousOperation("Invalid base64 encoded string.")
-    img = make_qr_code_image(text, qr_code_options=qr_code_options)
+    img = make_qr_code_image(data, qr_code_options=qr_code_options)
     return HttpResponse(content=img,
                         content_type='image/svg+xml' if qr_code_options.image_format == 'svg' else 'image/png')
 
 
 def get_qr_code_option_from_request(request) -> QRCodeOptions:
     request_query = request.GET.dict()
-    for key in ('text', 'token', 'cache_enabled'):
+    for key in ('data', 'token', 'cache_enabled'):
         request_query.pop(key, None)
     # Force typing for booleans.
     request_query['micro'] = int(request_query.get('micro', 0)) == 1