+++ qr_code/qrcode/maker.py
@@ -11,39 +11,44 @@ from qr_code.qrcode.serve import make_qr_code_url
 from qr_code.qrcode.utils import QRCodeOptions
 
 
-def make_qr(data: Any, qr_code_options: QRCodeOptions):
+def make_qr(data: Any, qr_code_options: QRCodeOptions, force_text: bool = True):
     """Creates a QR code that encodes the given `data` with the given `qr_code_options`.
 
+    :param str data: The data to encode
+    :param qr_code_options: Options to create and serialize the QR code.
+    :param bool force_text: Tells whether we want to force the `data` to be considered as text string and encoded in
+        byte mode.
     :rtype: segno.QRCode
     """
     # WARNING: For compatibility reasons, we still allow to pass __proxy__ class (lazy string). Moreover, it would be OK to pass anything that has __str__
     # attribute (e. g. class instance that handles phone numbers).
-    if isinstance(data, (bytes, int)):
-        return segno.make(data, **qr_code_options.kw_make())
-    else:
-        return segno.make(str(data), **qr_code_options.kw_make())
+    if force_text:
+        return segno.make(str(data), **qr_code_options.kw_make(), mode='byte')
+    return segno.make(data, **qr_code_options.kw_make())
 
 
-def make_qr_code_image(data: Any, qr_code_options: QRCodeOptions) -> bytes:
+def make_qr_code_image(data: Any, qr_code_options: QRCodeOptions, force_text: bool = True) -> bytes:
     """
-    Returns a bytes object representing a QR code image for the provided `data`.
+    Creates a bytes object representing a QR code image for the provided `data`.
 
     :param str data: The data to encode
     :param qr_code_options: Options to create and serialize the QR code.
+    :param bool force_text: Tells whether we want to force the `data` to be considered as text string and encoded in
+        byte mode.
     :rtype: bytes
     """
-    qr = make_qr(data, qr_code_options)
+    qr = make_qr(data, qr_code_options, force_text=force_text)
     out = io.BytesIO()
     qr.save(out, **qr_code_options.kw_save())
     return out.getvalue()
 
 
-def make_embedded_qr_code(data: Any, qr_code_options: QRCodeOptions) -> str:
+def make_embedded_qr_code(data: Any, qr_code_options: QRCodeOptions, force_text: bool = True) -> str:
     """
     Generates a <svg> or <img> tag representing the QR code for the given `data`.
     This tag can be embedded into an HTML document.
     """
-    qr = make_qr(data, qr_code_options)
+    qr = make_qr(data, qr_code_options, force_text=force_text)
     kw = qr_code_options.kw_save()
     # Pop the image format from the keywords since qr.png_data_uri / qr.svg_inline
     # set it automatically
@@ -54,12 +59,12 @@ def make_embedded_qr_code(data: Any, qr_code_options: QRCodeOptions) -> str:
     return mark_safe(qr.svg_inline(**kw))
 
 
-def make_qr_code_with_args(data: Any, qr_code_args: dict) -> str:
+def make_qr_code_with_args(data: Any, qr_code_args: dict, force_text: bool = True) -> str:
     options = _options_from_args(qr_code_args)
-    return make_embedded_qr_code(data, options)
+    return make_embedded_qr_code(data, options, force_text=force_text)
 
 
-def make_qr_code_url_with_args(data: Any, qr_code_args: dict) -> str:
+def make_qr_code_url_with_args(data: Any, qr_code_args: dict, force_text: bool = True) -> str:
     cache_enabled = qr_code_args.pop('cache_enabled', DEFAULT_CACHE_ENABLED)
     if not isinstance(cache_enabled, bool):
         cache_enabled = not cache_enabled == 'False'
@@ -67,7 +72,7 @@ def make_qr_code_url_with_args(data: Any, qr_code_args: dict) -> str:
     if not isinstance(url_signature_enabled, bool):
         url_signature_enabled = not url_signature_enabled == 'False'
     options = _options_from_args(qr_code_args)
-    return make_qr_code_url(data, options, cache_enabled=cache_enabled,
+    return make_qr_code_url(data, options, force_text=force_text, cache_enabled=cache_enabled,
                             url_signature_enabled=url_signature_enabled)