+++ qr_code/templatetags/qr_code.py
@@ -11,7 +11,7 @@ register = template.Library()
 
 
 def _make_app_qr_code_from_obj_or_kwargs(obj_or_kwargs, expected_cls, embedded: bool, qr_code_args: dict,
-                                         extra_qr_code_args: Optional[dict] = None) -> str:
+                                         extra_qr_code_args: Optional[dict] = None, force_text: bool = True) -> str:
     if isinstance(obj_or_kwargs, expected_cls):
         obj = obj_or_kwargs
     else:
@@ -21,9 +21,9 @@ def _make_app_qr_code_from_obj_or_kwargs(obj_or_kwargs, expected_cls, embedded: 
     if extra_qr_code_args:
         final_args.update(extra_qr_code_args)
     if embedded:
-        return make_qr_code_with_args(obj.make_qr_code_data(), qr_code_args=final_args)
+        return make_qr_code_with_args(obj.make_qr_code_data(), qr_code_args=final_args, force_text=force_text)
     else:
-        return make_qr_code_url_with_args(obj.make_qr_code_data(), qr_code_args=final_args)
+        return make_qr_code_url_with_args(obj.make_qr_code_data(), qr_code_args=final_args, force_text=force_text)
 
 
 def _make_google_maps_qr_code(embedded: bool, **kwargs) -> str:
@@ -49,8 +49,13 @@ def _make_geolocation_qr_code(embedded: bool, **kwargs) -> str:
 
 
 @register.simple_tag()
-def qr_from_text(text: Any, **kwargs) -> str:
-    return make_qr_code_with_args(text, qr_code_args=kwargs)
+def qr_from_text(text: str, **kwargs) -> str:
+    return make_qr_code_with_args(data=text, qr_code_args=kwargs)
+
+
+@register.simple_tag()
+def qr_from_data(data: Any, **kwargs) -> str:
+    return make_qr_code_with_args(data=data, qr_code_args=kwargs, force_text=False)
 
 
 @register.simple_tag()
@@ -104,15 +109,21 @@ def qr_for_wifi(wifi_config, **kwargs) -> str:
 def qr_for_epc(epc_data, **kwargs) -> str:
     extra = dict(
         error_correction='M',
-        boost_error=False
+        boost_error=False,
+        micro=False,
     )
     return _make_app_qr_code_from_obj_or_kwargs(epc_data, EpcData, embedded=True, qr_code_args=kwargs,
-                                                extra_qr_code_args=extra)
+                                                extra_qr_code_args=extra, force_text=False)
 
 
 @register.simple_tag()
-def qr_url_from_text(text: Any, **kwargs) -> str:
-    return make_qr_code_url_with_args(text, qr_code_args=kwargs)
+def qr_url_from_text(text: str, **kwargs) -> str:
+    return make_qr_code_url_with_args(data=text, qr_code_args=kwargs)
+
+
+@register.simple_tag()
+def qr_url_from_data(data: Any, **kwargs) -> str:
+    return make_qr_code_url_with_args(data=data, qr_code_args=kwargs, force_text=False)
 
 
 @register.simple_tag()
@@ -169,4 +180,4 @@ def qr_url_for_epc(epc_data, **kwargs) -> str:
         boost_error=False
     )
     return _make_app_qr_code_from_obj_or_kwargs(epc_data, EpcData, embedded=False, qr_code_args=kwargs,
-                                                extra_qr_code_args=extra)
+                                                extra_qr_code_args=extra, force_text=False)